% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callLPJ.R
\name{callLPJ}
\alias{callLPJ}
\title{A function to call the LPJ_GUESS modell}
\usage{
callLPJ(mainDir = NULL, runDir = NULL, template2 = NULL, mode = NULL)
}
\arguments{
\item{mainDir}{a character string indicating the path to LPJ}

\item{runDir}{a character string indicating the path to the runDirectory}

\item{template2}{a character string providing the  "specific" model template,
e.g, global_cf.ins or global_cru.ins. Provide only the file name, not the path.
If not provided, package templates will be used}

\item{mode}{a character string indicating whether using cru or cf data}
}
\value{
none
}
\description{
This function does a system call, passing a template to the LPJ
 model.
}
\note{
Based an older code of Istem Fer, Uni Potsdam
}
\examples{
\dontrun{
callLPJ("/home/LPJrun", "/home/LPJrun/runDirectory1", "global_cru.ins")
}
}
\author{
Florian Hartig, Ramiro Silveyra Gonzalez
}
\keyword{Rlpj}

