% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkParameters.R
\name{checkParameters}
\alias{checkParameters}
\title{A parameter list function}
\usage{
checkParameters(scale = NULL, parameterList = NULL)
}
\arguments{
\item{scale}{a character string indicating whether the model runs global or for europe.}

\item{parameterList}{a named list holding the parameter or combination or parameters
to be tested.}
}
\value{
a named list holding the values of the template parameter.
}
\description{
This function checks the provided parameter list against the
complete parameters list. If any parameter is not provided, the function will
add it to the parameter list and return a complete parameter list with the values
provided as input and the default values for those parameters that were not provided.
}
\examples{
\dontrun{
parameterList <- list(run_emax = 4, run_lambda_max = 1)
parameterList.checked <- checkParameters(scale= "global", parameterList)
}
}
\author{
Florian Hartig, Ramiro Silveyra Gonzalez, Maurizio Bagnara
}
\keyword{Rlpj}

