% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesInternal.R
\name{convertTS}
\alias{convertTS}
\title{The function to fill the single run object}
\usage{
convertTS(data = NULL)
}
\arguments{
\item{data}{TODO}
}
\value{
TODO
}
\description{
This function TODO
}

