% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSingleObject.R
\name{createSingleObject}
\alias{createSingleObject}
\title{The function to fill the single run object}
\usage{
createSingleObject(mainDir, typeList, settings)
}
\arguments{
\item{singleObject}{TODO}

\item{runDir}{TODO}

\item{outDir}{TODO}

\item{parameterList}{TODO}

\item{ID}{TODO}
}
\value{
TODO
}
\description{
This function TODO
}

