% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exit_mpi.R
\name{exit.mpi}
\alias{exit.mpi}
\title{The function to close MPI connection on cluster}
\usage{
exit.mpi()
}
\examples{
\dontrun{
exit.mpi()
 }
}
\author{
Florian Hartig, Ramiro Silveyra Gonzalez
}
\seealso{
\url{https://cran.r-project.org/web/packages/Rmpi/Rmpi.pdf}
}
\keyword{Rlpj}

