% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exitMPI.R
\name{exitMPI}
\alias{exitMPI}
\title{The function to close MPI connection on cluster.}
\usage{
exitMPI()
}
\description{
This function will close slaves and finalize mpi.
}
\details{
The exitMPI should be use when working on MPI clusters. It must be
called at the end of you script. Be aware that when mpi is exited, it will be
no longer possible to work on MPI clusters withouht relaunching R.
The function is a wrapper of mpi.finalize from the Rmpi package. Check the
package manual for futher advise on using mpi.finalize and mpi.quit.
}
\examples{
\dontrun{
exitMPI()
 }
}
\author{
Ramiro Silveyra Gonzalez
}
\seealso{
\url{https://cran.r-project.org/web/packages/Rmpi/Rmpi.pdf}
}
\keyword{Rlpj}

