% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesInternal.R
\name{fillSingleObject}
\alias{fillSingleObject}
\title{The function to fill the single run object}
\usage{
fillSingleObject(singleObject, runDir, outDir, parameterList, ID)
}
\arguments{
\item{singleObject}{TODO}

\item{runDir}{TODO}

\item{outDir}{TODO}

\item{parameterList}{TODO}

\item{ID}{TODO}
}
\value{
TODO
}
\description{
This function TODO
}

