% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{A function to process LPJ outputs}
\usage{
getData(typeList = NULL, outDir = NULL, runInfo = NULL,
  processing = TRUE)
}
\arguments{
\item{typeList}{a character vector with the outputs to be analyzed.
Default value is all outputs that were cretaed}

\item{outDir}{a character string indicating path to the output directory}

\item{runInfo, }{a named list with the information of the LPJ run.
The runInfo it will be stored by the function as RData along with
the processed outputs of the model}

\item{processing}{a boolean indicating whether output files will be turned into time series}
}
\value{
the processed data returned in a S4 Class: LPJData Class
}
\description{
This function reads the ASCII outputs produced by LPJ-GUESS.
It takes a list of output types (typeList) and finds them in the specified path.
The data is stored as a matrix with the shape of a zoo time series and
packed in a data class object: LPJData.
}
\note{
Based on an older code of Joerg Steinkamp
}
\examples{
\dontrun{
LPJout <- getData( typeList = c("aaet",  "cflux","lai", "nflux"),
          "~/path/to/output/files", runInfo = list(parameter1 = 0.5, grid = 1))
}
}
\author{
Florian Hartig, Ramiro Silveyra Gonzalez, Maurizio Bagnara
}
\seealso{
\url{https://cran.r-project.org/web/packages/zoo/zoo.pdf}
}

