% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesExported.R
\name{getParameterList}
\alias{getParameterList}
\title{A get paremeter list function}
\usage{
getParameterList(scale = NULL)
}
\arguments{
\item{scale}{a character string indicating whether the parameters are for the global or europe template}
}
\value{
a list with parameter names and their values
}
\description{
This function returns the LPJ-GUESS parameter list required for the writeTemplate function.
It also allows users to see the default values of the templates.
}
\examples{
\dontrun{
parameterList <- getParameterList("global")
}
}

