% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRunInfo.R
\name{getRunInfo}
\alias{getRunInfo}
\title{A get runInfo data}
\usage{
getRunInfo(x, parameters = F)
}
\arguments{
\item{x}{a character string indicating the absolute path to the runInfoDir folder or
a R object produced by runLPJ}

\item{parameters}{a character boolean to specify whether to return the parameters
instead of the LPJData objects}
}
\description{
This function reads the runLPj outputs stored in the runInfoDir
and returns as a list, with the same structure as the runLPJ outputs. Additionally,
allows to retrieve the parameters from the runInfoDir files or the runLPJ outputs.
}
\examples{
\dontrun{
# recover the data from the runInfo folder
result <- getRunInfo("/some/absolute/path/runInfo_2016_08_11_121507")

# recover only the paramaters from the runInfo folder
parameters <- getRunInfo("/some/absolute/path/runInfo_2016_08_11_121507", parameters =T)

# recover only the parameters from the result
parameters <- getRunInfo(result, parameters =T)

}
}
\author{
Ramiro Silveyra Gonzalez
}

