% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJRunObject.R, R/LPJSetupClass.R
\docType{methods}
\name{TODO}
\alias{TODO}
\title{TODO}
\usage{
\S4method{initialize}{LPJSetup}(.Object, clusterType, numCores, mainDir, runDir,
  outDir)

\S4method{initialize}{LPJSetup}(.Object, clusterType, numCores, mainDir, runDir,
  outDir)
}

