% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJDataClass.R
\docType{class}
\name{LPJData-class}
\alias{LPJData-class}
\title{A LPJ-GUESS data class object#'}
\value{
a data class object
}
\description{
This is a data class object
}
\note{
Based an older code of Istem Fer, Uni Potsdam
}
\author{
Florian Hartig, Ramiro Silveyra Gonzalez
}
\keyword{Rlpj}

