% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLPJData.R
\name{plotLPJData}
\alias{plotLPJData}
\title{A plot function for LPJdata objects}
\usage{
plotLPJData(data = NULL, typeList = NULL, outDir = NULL,
  save.plots = FALSE, prefix = "")
}
\arguments{
\item{data}{a list of data contained in the LPJ data object.}

\item{typeList}{a character vector with the outputs to be plotted}

\item{outDir}{a character string indicating the folder where the plots will be
saved, if save.plot set to TRUE}

\item{save.plots}{a boolean indicating whether the plots are saved in the outDir}

\item{prefix}{a character string specifying the prefix to be added to the plots files.
Only relevant if saving plots.}
}
\value{
plots for data types included in typeList
}
\description{
This function reads data from a LPJ data object, turns it into
a time series with zoo and plots the variables against time.
Plots are saved in the output folder, if the save.plots boolean is set to TRUE.
Data in the data object is already a time series object, but stored as a matrix.
}
\examples{
\dontrun{
plotLPJData(dataList = list( aaet = aaet), typeList = c("aaet"),
 outDir = "/runDir/outDir", save.plots = FALSE)
}
}
\author{
Florian Hartig, Ramiro Silveyra Gonzalez
}
\seealso{
\url{https://cran.r-project.org/web/packages/zoo/zoo.pdf}
}
\keyword{Rlpj}

