% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLPJparallel.R
\name{runLPJParallel}
\alias{runLPJParallel}
\title{The function to run the LPJ in parallel}
\usage{
runLPJParallel(setupObject, plot.data = FALSE, save.plots = FALSE,
  parameterList = NULL, file.co2 = NULL, file.cru = NULL,
  file.cru.misc = NULL, file.ndep = NULL, file.temp = NULL,
  file.prec = NULL, file.insol = NULL)
}
\arguments{
\item{setupObject}{a named list created by the setupParallel function containing
the following: mainDir, template1, template2, gridList, clusterType, numCores,
 typeList, mode and scale.}

\item{plot.data}{a boolean indicating whether the ouput data will be plotted.}

\item{save.plots}{a boolean indicating whether the plots will be saved.}

\item{parameterList}{a named list containing the parameters to be calibrated}

\item{file.co2}{a character string providing the absolute path to the C02 input file}

\item{file.cru}{a character string providing the absolute path to the cur? input file}

\item{file.cru.misc}{a character string providing the absolute path to the cru?
input file}

\item{file.ndep}{a character string providing the absolute path to the nitrogen
deposition input file}

\item{file.temp}{a character string providing the absolute path to the temperature
input file}

\item{file.prec}{a character string providing the absolute path to the
precipitation input file}

\item{file.insol}{a character string providing the absolute path to the
insolation input file}
}
\value{
a list holding the outputs of the runLPJwrapper (see function help)
}
\description{
This function reads the setup parallel object and creates a
 cluster to which it submits the model wrapper function with its respective
 parameters.
}
\details{
The runLPJparallel assumes the existence of a folder containing all
the inputs data and templates for LPJ-GUESS and a directory structure for
storing inputs and outputs of each single run. The setupParallel function is
ought to be run before calling the runLPJparallel.
Running the LPJ parallel involves two steps. First, to create a parallel
setup (setupParallel function), and second, to actually run in parallel the model
(runLPJparallel function).  The parallelization requires the packages snow and
 also the Rmpi package, if you aim at using a MPI cluster.
}
\examples{
\dontrun{
# You need to specify the absolute path of each input file:
file.co2<-"/some/absolute/path/crudata/co2_1901-2013_FAKE.txt"
file.cru <- "/some/absolute/path/crudata/cru_1901_2006.bin"
file.cru.misc <- "/some/absolute/path/crudata/cru_1901_2006misc.bin"
file.ndep <- "/some/absolute/path/crudata/GlobalNitrogenDeposition.bin"

# If you are using the global_cf.ins file you need to specify the site
# specific input files as well
file.temp <- "/some/absolute/path/cfdata/temp.nc"
file.prec <- "/some/absolute/path/cfdata/prec.nc"
file.insol <- "/some/absolute/path/cfdata/rad.nc"

# Create some paramaters to test modell.
# Number of runs is proportional to number of parameter set you are testing
parameterDefault <- list (run_emax = NULL)

# Test 6 different values for emax.
par <- seq(1,5, len = 6)
# Create the list object with the parameters
parameterList <- vector("list", length(par))
for (i in 1:length(par)) {
   parameterDefault$run_emax <- par[i]
   parameterList[[i]] <- parameterDefault
 }

# Call setupParallel
setupObject  <- setupParallel(3, "SOCK", "global", "cf", gridList = "gridlist_geb.txt",
                             mainDir = "/some/absolute/path/mainDir")

# Call the runLPjParallel
result <- runLPJParallel(setupObject , plot.data = FALSE, save.plots = FALSE,
                         parameterList=parameterList, file.co2, file.cru,
                         file.cru.misc, file.ndep, file.temp, file.prec, file.insol)

   Checking conditions
   Reading the parallel object structure
   Creating the single run objects....1....2....3....4....5....6
   Creating a SOCK cluster with 3 cores
   Sending tasks to the cores

   Processing ended!

str(result,1)
   List of 6
   $ :Formal class 'LPJData' [package "Rlpj"] with 2 slots
   $ :Formal class 'LPJData' [package "Rlpj"] with 2 slots
   $ :Formal class 'LPJData' [package "Rlpj"] with 2 slots
   $ :Formal class 'LPJData' [package "Rlpj"] with 2 slots
   $ :Formal class 'LPJData' [package "Rlpj"] with 2 slots
   $ :Formal class 'LPJData' [package "Rlpj"] with 2 slots

str(result[[1]], 2 )
   Formal class 'LPJData' [package "Rlpj"] with 2 slots
   ..@ runInfo  :List of 15
   ..@ dataTypes:List of 39


 }
}
\author{
Florian Hartig, Ramiro Silveyra Gonzalez
}
\seealso{
\url{https://cran.r-project.org/web/packages/Rmpi/Rmpi.pdf},
 \url{https://cran.r-project.org/web/packages/snow/snow.pdf}
}
\keyword{Rlpj}

