% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLPJwrapper.R
\name{runLPJwrapper}
\alias{runLPJwrapper}
\title{A wrapper to run the LPJ-GUESS and processed the ouputs}
\usage{
runLPJwrapper(runObject)
}
\arguments{
\item{runObject}{a named list created by the runLPJParallel function containing
the following: mainDir, template1, template2, gridList, runDir, outDir, mode,
scale, typeList, runID,  gridFilename, plot.data and save.plots}
}
\value{
a a data class object with the slots runInfo and dataTypes. The runInfo
slot contains the provided run information and parameters; the dataTypes holds
the processed output data from the run.
}
\description{
This function modifies the lpj template with the desired parameter
 values and then does a system call for running the lpj modell.
 The modell outputs are processed, plotted (if set to TRUE)
}
\note{
Based an older code of Istem Fer, Uni Potsdam
}
\examples{
\dontrun{
runLPJwrapper(runObject)
}
}
\author{
Florian Hartig, Ramiro Silveyra Gonzalez, Maurizio Bagnara
}
\keyword{Rlpj}

