% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupLPJParallel.R
\name{setupLPJParallel}
\alias{setupLPJParallel}
\title{The function to create a setup for parallel runs of the LPJ-GUESS}
\usage{
setupLPJParallel(numCores = NULL, clusterType = NULL, mainDir = NULL)
}
\arguments{
\item{numCores}{a integer specifying number of cores of the cluster}

\item{clusterType}{a character string indicating the type of cluster to be created. If running in bwHPC,
type must be "MPI". If running in personal computer, type must be "SOCK".}

\item{mainDir}{a character string indicating the path to the directory where
all input data and template are located and in which the function will create
the directory structure for the outputs}
}
\value{
a setup object or named list containing the setup parameters to run
the LPJ in parallel
}
\description{
This function first creates a setup for running the LPJ in parallel
 based on the provided input parameters. The function assumes a specific initial
 configuration. A folder (mainDir) containing the input data, the templates,
 and link to the model must exist. A directory structure for the outputs will
 be build wihtin this folder by the function.
}
\examples{
\dontrun{
mainDir <- "/some/absolute/path/mainDir"
list.files(mainDir)
[1] "guess" or "guesscmd.exe"  # link to the model executable
[2] "gridlist.txt"      # list of gridcells
[3] "global.ins"        # template1 (optional)
[4] "global_cru.ins"    # template2 (optional)

mySetup <- setupLPJParallel(numCores= 3, clusterType = "SOCK", mainDir=mainDir)
str(mySetup, 1)
     $ clusterType: chr "SOCK"
     $ numCores   : num 3
     $ runDir     : chr [1:3] "/some/absolute/path/mainDir/runDirectory1"
                              "/some/absolute/path/mainDir/runDirectory2"
                               "/some/absolute/path/mainDir/runDirectory3"
     $ outDir     : chr [1:3] "/some/absolute/path/mainDir/runDirectory1/outDirectory1"
                              "/some/absolute/path/mainDir/runDirectory2/outDirectory2"
                               "/some/absolute/path/mainDir/runDirectory3/outDirectory3"
}
}
\author{
Florian Hartig, Ramiro Silveyra Gonzalez, Maurizio Bagnara
}
\seealso{
\url{https://cran.r-project.org/web/packages/Rmpi/Rmpi.pdf},
 \url{https://cran.r-project.org/web/packages/snow/snow.pdf}
}
\keyword{Rlpj}

