% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupParallel.R
\name{setupParallel}
\alias{setupParallel}
\title{The function to create a setup for parallel runs of the LPJ-GUESS}
\usage{
setupParallel(numCores = NULL, clusterType = NULL, scale = NULL,
  mode = NULL, template1 = NULL, template2 = NULL, gridList = NULL,
  mainDir = NULL, typeList = NULL)
}
\arguments{
\item{numCores}{a integer specifying number of cores of the cluster}

\item{clusterType}{a character string indicating the type of cluster to be created. If running in bwHPC,
type must be "MPI". If running in personal computer, type must be "SOCK".}

\item{scale}{a character string indicating whether the model runs global or
for europe.}

\item{mode}{a character string indicating whether using cru or cf data}

\item{template1}{a character string providing the general model template,
e.g, global.ins. Provide only the file name, not the path.
If not provided, package templates will be used.}

\item{template2}{a character string providing the  "specific" model template,
e.g, global_cf.ins or global_cru.ins. Provide only the file name, not the path.
If not provided, package templates will be used.}

\item{gridList}{a character string providing the name of the text file with
the grids to be included in the model, e.g gridlist.txt.
Provide only the file name, not the path.}

\item{mainDir}{a character string indicating the path to the directory where
all input data and template are located and in which the function will create
the directory structure for the outputs}

\item{typeList}{a character vector with the outputs to be analyzed.
Default value is all outputs.}
}
\value{
a setup object or named list containing the setup paremeters to run
the LPJ in parallel
}
\description{
This function first creates a setup for running the LPJ in parallel
 based o the provided input parameters, and then runs the model in parallel.
 The function assumes a specific initial configuration. A folder (mainDir)
 containing the input data, the templates, and link to the model must exist.
 A directory structure for the outputs will be build wihtin this folder by
 the function.
}
\examples{
\dontrun{
mainDir <- "~/lpjRun"
list.files(mainDir)
[1] "guess" or "guesscmd.exe"  # link to the model executable
[2] "gridlist.txt"      # list of gridcells
[3] "global.ins"        # template1 (optional)
[4] "global_cru.ins"    # template2 (optional)

setupObject <- setupParallel(numCores= 10, clusterType = "SOCK",
                             gridList= "gridlist.txt", mainDir=mainDir,typeList =NULL,
                             scale = "europe", mode = "cf")

     Output type list has not been provided.
     Setting type list to default values.

     Using package template (template 1).
     Saving package template in the mainDir.

     Using package template (template 2).
     Saving package template in the mainDir.

     Creating the foder structure....1....2....3....4....5....6....7....8....9....10

str(setupObject)
     List of 9
     $ mainDir    : chr "/home/lpjRun"
     $ template1  : chr "europe.ins"
     $ template2  : chr "europe_cf.ins"
     $ gridList   : chr "gridlist.txt"
     $ clusterType: chr "SOCK"
     $ numCores   : num 10
     $ typeList   : chr [1:39] "aaet" "agpp" "aiso" "amon" ...
     $ mode       : chr "cf"
     $ scale      : chr "europe"
}
}
\author{
Florian Hartig, Ramiro Silveyra Gonzalez, Maurizio Bagnara
}
\seealso{
\url{https://cran.r-project.org/web/packages/Rmpi/Rmpi.pdf},
 \url{https://cran.r-project.org/web/packages/snow/snow.pdf}
}
\keyword{Rlpj}

