% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLPJout.R
\name{writeLPJout}
\alias{writeLPJout}
\title{A write LPj outputs function}
\usage{
writeLPJout(runInfo = NULL, typeList = NULL, outDir = NULL)
}
\arguments{
\item{runInfo}{a list of matrices contained in the LPJ data object.}

\item{typeList}{a character vector with the outputs to be written out.}

\item{outDir}{a character string indicating the folder where the plots will be
saved.}
}
\description{
This function will allow you write out the files contained in
runInfo.Rdata.
}
\examples{
\dontrun{
plotData(dataList = list( aaet = aaet), typeList = c("aaet"),
 outDir = "/runDir/outDir", save.plots = FALSE)
}
}
\author{
Florian Hartig, Ramiro Silveyra Gonzalez
}
\keyword{Rlpj}

