% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesInternal.R
\name{convertTS}
\alias{convertTS}
\title{The function to convert data into time series}
\usage{
convertTS(data = NULL)
}
\arguments{
\item{data}{a data frame containing the LPJ model outputs}
}
\value{
a montly or yearly zoo time seires object
}
\description{
This function transform the data into a zoo time series. The
variabels Lon, Lat and Year are dropped.
}

