% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exit_mpi.R
\name{exit.mpi}
\alias{exit.mpi}
\title{The function to close MPI connection on cluster.}
\usage{
exit.mpi()
}
\description{
This function will close slaves and finalize mpi.
}
\details{
The exit.mpi should be use when working on MPI clusters. It must be
called at the end of you script. Be aware that when mpi is exited, it will be
no longer possible to work on MPI clusters withouht relaunching R.
}
\examples{
\dontrun{
exit.mpi()
 }
}
\author{
Ramiro Silveyra Gonzalez
}
\seealso{
\url{https://cran.r-project.org/web/packages/Rmpi/Rmpi.pdf}
}
\keyword{Rlpj}

