% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesInternal.R
\name{fillSingleObject}
\alias{fillSingleObject}
\title{The function to fill the single run object}
\usage{
fillSingleObject(singleObject, runDir, outDir, parameterList, ID)
}
\arguments{
\item{singleObject}{a named list}

\item{runDir}{a character string indicating path to the run directory}

\item{outDir}{a character string indicating path to the output directory}

\item{parameterList}{a named list containing the parameters to be calibrated}

\item{ID}{an integer indicating the run number}
}
\value{
the filled singleObject
}
\description{
This function completes the sinlge object with the required parameters
}

