% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{A function to process LPJ outputs}
\usage{
getData(typeList = NULL, outDir = NULL, runInfo = NULL,
  lon.extent = c(-180, 180), lat.extent = c(-90, 90),
  area.weighted = FALSE, year.offset = 0)
}
\arguments{
\item{typeList}{a character vector with the outputs to be analyzed.
Default value is all outputs that were cretaed}

\item{outDir}{a character string indicating path to the output directory}

\item{runInfo, }{a named list with the information of the LPJ run.
The runInfo it will be stored by the function as RData along with
the processed outputs of the model}

\item{lon.extent}{a numeric vector containing the min and max values used
for west-east extent(default -180 to 180)}

\item{lat.extent}{a numeric vector containing the max and min values used
for north-south extent (default 90 to -90).}

\item{area.weighted}{a boolean indicating whether the gridcells should be
weighted by their size (regular grids only, default FALSE).}

\item{year.offset}{a integer indicating the value to be added to the 'Year'
column in the LPJ-GUESS output.}
}
\value{
the processed data returned in a S4 Class: LPJData Class
}
\description{
This function reads the ASCII outputs produced by LPJ-GUESS.
It takes a list of output types (typeList) and finds them in the specified path.
The data is stored as a matrix with the shape of a zoo time series and
packed in a data class object: LPJData.
}
\note{
Based on an older code of Joerg Steinkamp
}
\examples{
\dontrun{
LPJout <- getData( typeList = c("aaet",  "cflux","lai", "nflux"),
          "~/path/to/output/files", runInfo = list(parameter1 = 0.5, grid = 1))
}
}
\author{
Florian Hartig, Ramiro Silveyra Gonzalez, Maurizio Bagnara
}
\seealso{
\url{https://cran.r-project.org/web/packages/zoo/zoo.pdf}
}

