% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTemplate.R
\name{getTemplate}
\alias{getTemplate}
\title{A get template function}
\usage{
getTemplate(type = NULL, outputDir = NULL)
}
\arguments{
\item{type}{a character string indicating the template name: global,
global_cf, global_cru, europe, europe_cf, europe_cru}

\item{outputDir}{a character string indicating path to the output directory  (optional)}
}
\value{
a template object or template.ins file in the specified folder
}
\description{
This function returns the LPJ-GUESS templates.
}
\examples{
\dontrun{
template <- getTemplate("global.ins")
getTemplate("global.ins", "/home/LPJTemplates/")
}
}
\author{
Ramiro Silveyra Gonzalez, Maurizio Bagnara, Florian Hartig
}
\keyword{Rlpj}

