% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJDataClass.R
\docType{class}
\name{LPJData-class}
\alias{LPJData-class}
\title{A LPJ-GUESS data class object}
\value{
a data class object with the slots runInfo and dataTypes
}
\description{
This is a data class object to store both the LPJ-GUESS outputs
and the information that was passed to LPJ-GUESS.
}
\keyword{#'}
\keyword{@author}
\keyword{Gonzalez}
\keyword{Ramiro}
\keyword{Rlpj}
\keyword{Silveyra}

