% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJRunObject.R, R/LPJSetupClass.R
\docType{class}
\name{LPJSetup-class}
\alias{LPJSetup-class}
\title{A LPJ-GUESS data class object}
\value{
a data class object with the slots runInfo and dataTypes

a data class object with the slots runInfo and dataTypes
}
\description{
This is a data class object to store both the LPJ-GUESS outputs
and the information that was passed to LPJ-GUESS.

This is a data class object to store both the LPJ-GUESS outputs
and the information that was passed to LPJ-GUESS.
}
\author{
Ramiro Silveyra Gonzalez

Ramiro Silveyra Gonzalez
}
\keyword{Rlpj}

