% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLPJParallel.R
\name{MPISapply}
\alias{MPISapply}
\title{The function to run the LPJ in parallel}
\usage{
MPISapply(numcores, runParameters)
}
\arguments{
\item{numcores}{a integer specifying number of cores of the cluster}

\item{runParameters}{a list of lists, each list containing the following information:
mainDir, template1, template2, gridList, runDir, outDir, mode, scale,
 typeList, parameterList, runID and gridFilename}
}
\description{
This function tells each core which are its tasks and makes the
core run the runLPJwrapper on them using sapply
}
\note{
based on lapplys from M. T. Morgan (mtmorgan@fhcrc.org) (Parallel R)
}
\examples{
\dontrun{#'
result <- MPISapply(numcores = 6, runParameters = runParameters)
}
}
\author{
Ramiro Silveyra Gonzalez
}
\keyword{Rlpj}

