% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLPJData.R
\name{plotLPJData}
\alias{plotLPJData}
\title{A plot function for LPJData objects}
\usage{
plotLPJData(x = NULL, typeList = NULL, outDir = NULL,
  save.plots = FALSE, prefix = "")
}
\arguments{
\item{x}{a LPJData object.}

\item{typeList}{a character vector with the outputs to be plotted}

\item{outDir}{a character string indicating the folder where the plots will be
saved, if save.plot set to TRUE}

\item{save.plots}{a boolean indicating whether the plots are saved in the outDir}

\item{prefix}{a character string specifying the prefix to be added to the plots files.
Only relevant if saving plots}
}
\value{
plots for data types included in typeList. The grid cells will be plotted independently
}
\description{
This function reads data from a LPJData object and plots the
 variables against time. Plots are saved in the output folder, if the save.plots is set to TRUE.
}
\examples{
\dontrun{
plotLPJData(data, typeList = c("aaet", "lai"),
 outDir = "/runDir/outDir", save.plots = FALSE)
}
}
\author{
Florian Hartig, Ramiro Silveyra Gonzalez
}
\seealso{
\url{https://cran.r-project.org/web/packages/zoo/zoo.pdf}
}
\keyword{Rlpj}

