% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLPJparallel.R
\name{runLPJParallel}
\alias{runLPJParallel}
\title{The function to run the LPJ in parallel}
\usage{
runLPJParallel(setupObject, plot.data = FALSE, save.plots = FALSE,
  parameterList = NULL, file.co2 = NULL, file.cru = NULL,
  file.cru.misc = NULL, file.ndep = NULL, file.temp = NULL,
  file.prec = NULL, file.insol = NULL)
}
\arguments{
\item{setupObject}{a named list created by the setupParallel function containing
the following: mainDir, template1, template2, gridList, clusterType, numCores,
 typeList, mode and scale.}

\item{plot.data}{a boolean indicating whether the ouput data will be plotted.}

\item{save.plots}{a boolean indicating whether the plots will be saved.}

\item{parameterList}{a named list containing the parameters to be calibrated}

\item{file.co2}{a character string providing the absolute path to the C02 input file}

\item{file.cru}{a character string providing the absolute path to the cur? input file}

\item{file.cru.misc}{a character string providing the absolute path to the cru?
input file}

\item{file.ndep}{a character string providing the absolute path to the nitrogen
deposition input file}

\item{file.temp}{a character string providing the absolute path to the temperature
input file}

\item{file.prec}{a character string providing the absolute path to the
precipitation input file}

\item{file.insol}{a character string providing the absolute path to the
insolation input file}
}
\value{
a list holding the outputs of the runLPJwrapper (see function help)
}
\description{
This function reads the setup parallel object and creates a
 cluster to which it submits the model wrapper function with its respective
 parameters.
}
\details{
The runLPJparallel assumes the existence of a folder containing all
the inputs data and templates for LPJ-GUESS and a directory structure for
storing inputs and outputs of each single run. The setupParallel function is
ought to be run before calling the runLPJparallel.
}
\examples{
\dontrun{
result <- runLPJParallel(setupObject , plot.data = FALSE, save.plots = FALSE,
parameterList=parameterList, file.co2, file.cru, file.cru.misc, file.ndep,
file.temp, file.prec, file.insol)
 }
}
\author{
Florian Hartig, Ramiro Silveyra Gonzalez
}
\seealso{
\url{https://cran.r-project.org/web/packages/Rmpi/Rmpi.pdf},
 \url{https://cran.r-project.org/web/packages/snow/snow.pdf}
}
\keyword{Rlpj}

